api/requirements.txt
# Azure Functions Core
azure-functions

# Azure Identity for Authentication
azure-identity

# Microsoft Fabric Dependencies
semantic-link
semantic-link-sempy

fabric-data-agent-sdk
